third_party/WebKit/LayoutTests/fast/dnd/file-drop-on-webkitdropzone-element.html

<!doctype html>
<meta charset="utf-8" />
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="resources/copy-data-transfer.js"></script>
<style>
#dropzone {
  display: block;
  border: 1px solid black;
  width: 200px;
  height: 200px;
}
</style>

<p>
  Please download <a download href="resources/dragged-file.txt">this file</a>,
  and drag it into the box below.
</p>

<div id="dropzone" webkitdropzone="copy file:text/plain">
  Drop Here
</div>

<script>
'use strict';

const dropZone = document.querySelector('#dropzone');

const dropPromise = new Promise((resolve, reject) => {
  dropZone.ondrop = event => {
    event.preventDefault();  // Needed to prevent drop navigation.
    resolve(copyDataTransfer(event.dataTransfer));
  };
});

const clientRect = dropZone.getBoundingClientRect();
if (window.eventSender) {
  eventSender.beginDragWithFiles(['resources/dragged-file.txt']);
  const centerX = (clientRect.left + clientRect.right) / 2;
  const centerY = (clientRect.top + clientRect.bottom) / 2;
  eventSender.mouseMoveTo(centerX, centerY);
  eventSender.mouseUp();
}

promise_test(() => {
  return dropPromise.then((dataTransfer) => {
    assert_array_equals(dataTransfer.types, ['Files']);
  });
}, 'DataTransfer.types in drop');

promise_test(() => {
  return dropPromise.then((dataTransfer) => {
    assert_object_equals(dataTransfer.data, {Files: ''});
  });
}, 'DataTransfer.getData() return values in drop');

promise_test(() => {
  return dropPromise.then((dataTransfer) => {
    assert_equals(
        dataTransfer.files.length, 1,
        'DataTransfer.files should have one element');
    const file = dataTransfer.files[0];
    assert_equals(file.file.name, 'dragged-file.txt');
    assert_equals(file.file.type, 'text/plain');
    assert_equals(file.file.size, 21);
    assert_equals(file.data, 'The test has FAILED.\n');
  });
}, 'DataTransfer.files in drop');

promise_test(() => {
  return dropPromise.then((dataTransfer) => {
    const file = dataTransfer.files[0].file;
    assert_true(
        file instanceof File,
        'DataTransfer.files[0] should be a File instance');
    assert_equals(
        file.name, 'dragged-file.txt',
        "File.name should reflect the dropped file's name");
    assert_equals(
        file.type, 'text/plain',
        "File.type should reflect the dropped file's MIME type");
    assert_equals(
        file.size, 21,
        "File.type should reflect the dropped file's size");
  });
}, 'DataTransfer.files[0] File metadata in drop');

promise_test(() => {
  return dropPromise.then((dataTransfer) => {
    assert_equals(dataTransfer.files[0].data, 'The test has FAILED.\n');
  });
}, 'Using FileReader to read DataTransfer.files[0] in drop');

promise_test(() => {
  return dropPromise.then((dataTransfer) => {
    assert_equals(
        dataTransfer.items.length, 1,
        'DataTransfer.items should have 1 element');
    const item = dataTransfer.items[0];
    assert_equals(
        item.kind, 'file',
        'DataTransferItem.kind should indicate that a file was dropped');
    assert_equals(
        item.type, 'text/plain',  // Firefox returns application/x-moz-file
        "DataTransferItem.type should reflect the dropped file's MIME type");
  });
}, 'DataTransfer.items in drop');

promise_test(() => {
  return dropPromise.then((dataTransfer) => {
    assert_equals(
        dataTransfer.items[0].error, undefined,
        'Reading the file in DataTransfer.items[0] should not throw');

    const file = dataTransfer.items[0].file;
    assert_true(
        file instanceof File,
        'DataTransfer.items[0].getAsFile() should return a File instance');
    assert_equals(
        file.name, 'dragged-file.txt',
        "File.name should reflect the dropped file's name");
    assert_equals(
        file.type, 'text/plain',
        "File.type should reflect the dropped file's MIME type");
    assert_equals(
        file.size, 21,
        "File.type should reflect the dropped file's size");
  });
}, 'DataTransfer.items[0].getAsFile() File metadata in drop');

promise_test(() => {
  return dropPromise.then((dataTransfer) => {
    assert_equals(dataTransfer.items[0].data, 'The test has FAILED.\n');
  });
}, 'Using FileReader to read DataTransfer.items[0].getAsFile() in drop');

</script>
